Move recording UI logic and css into separate file
