--- src/css/main.css
@@ -1,2 +1,0 @@
-@import "./poppins.css";
-@import "./welcome.css";
