+++ src/index.html
@@ -91,11 +91,7 @@
                 Start call
               </button>
               <button
-                class="
-                  welcome-page-button
-                  welcome-page-button-with-icon
-                  welcome-page-gutter-top
-                "
+                class="welcome-page-button welcome-page-button-with-icon welcome-page-gutter-top"
                 id="copy_link"
               >
                 <div class="copy-icon"></div>
@@ -103,11 +99,7 @@
               </button>
 
               <div
-                class="
-                  welcome-page-button
-                  welcome-page-button-with-icon
-                  welcome-page-button-loading
-                "
+                class="welcome-page-button welcome-page-button-with-icon welcome-page-button-loading"
                 id="enter_room_loading"
               >
                 <div class="spinner"></div>
@@ -142,7 +134,7 @@
               </div>
             </div>
             <div class="section recordings" id="recordings">
-              <p class="recordings-header">Saved Recordings</p>
+              <p class="recordings-header">Your recorded calls</p>
             </div>
             <div class="section subscribe" id="subscribe">
               <p class="subscribe-text">
